requirements.txt
torch==2.1.0+cu121
torchvision==0.16.0+cu121
numpy
pillow
matplotlib
tqdm
